[REFACTOR] Purchasement 클래스명을 PurchaseAmount로 변경합니다
구입 금액의 의미를 명시하기 위해 이름을 변경하였습니다.
